# apps/kajovo-hotel-api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_issue_photos_api_v1_issues__issue_id__photos_post": {
        "properties": {
          "photos": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Photos",
            "type": "array"
          }
        },
        "required": [
          "photos"
        ],
        "title": "Body_upload_issue_photos_api_v1_issues__issue_id__photos_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueCreate": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "location": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority",
            "default": "medium"
          },
          "room_number": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Number"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus",
            "default": "new"
          },
          "title": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "location"
        ],
        "title": "IssueCreate",
        "type": "object"
      },
      "IssuePriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "IssuePriority",
        "type": "string"
      },
      "IssueRead": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "in_progress_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Progress At"
          },
          "location": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/MediaPhotoRead"
            },
            "title": "Photos",
            "type": "array"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority",
            "default": "medium"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "room_number": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Number"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus",
            "default": "new"
          },
          "title": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "title",
          "location",
          "id",
          "in_progress_at",
          "resolved_at",
          "closed_at",
          "created_at",
          "updated_at"
        ],
        "title": "IssueRead",
        "type": "object"
      },
      "IssueStatus": {
        "enum": [
          "new",
          "in_progress",
          "resolved",
          "closed"
        ],
        "title": "IssueStatus",
        "type": "string"
      },
      "IssueUpdate": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssuePriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "room_number": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Number"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "IssueUpdate",
        "type": "object"
      },
...
      "MediaPhotoRead": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "thumb_path": {
            "title": "Thumb Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sort_order",
          "mime_type",
          "size_bytes",
          "file_path",
          "thumb_path",
          "created_at"
        ],
        "title": "MediaPhotoRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "K\u00e1jovoHotel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/issues": {
      "get": {
        "operationId": "list_issues_api_v1_issues_get",
        "parameters": [
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssuePriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "room_number",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room Number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueRead"
                  },
                  "title": "Response List Issues Api V1 Issues Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Issues",
        "tags": [
          "issues"
        ]
      },
      "post": {
        "operationId": "create_issue_api_v1_issues_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Issue",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/v1/issues/{issue_id}": {
      "delete": {
        "operationId": "delete_issue_api_v1_issues__issue_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Issue",
        "tags": [
          "issues"
        ]
      },
      "get": {
        "operationId": "get_issue_api_v1_issues__issue_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Issue",
        "tags": [
          "issues"
        ]
      },
      "put": {
        "operationId": "update_issue_api_v1_issues__issue_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Issue",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/photos": {
      "get": {
        "operationId": "list_issue_photos_api_v1_issues__issue_id__photos_get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaPhotoRead"
                  },
                  "title": "Response List Issue Photos Api V1 Issues  Issue Id  Photos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Issue Photos",
        "tags": [
          "issues"
        ]
      },
      "post": {
        "operationId": "upload_issue_photos_api_v1_issues__issue_id__photos_post",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_issue_photos_api_v1_issues__issue_id__photos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaPhotoRead"
                  },
                  "title": "Response Upload Issue Photos Api V1 Issues  Issue Id  Photos Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Issue Photos",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/photos/{photo_id}/{kind}": {
      "get": {
        "operationId": "get_issue_photo_api_v1_issues__issue_id__photos__photo_id___kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Issue Photo",
        "tags": [
          "issues"
        ]
      }
    },
...
  }
}